+++ package.json
@@ -4,6 +4,7 @@
   "repository": "git@github.com:TryGhost/TrafficAnalytics.git",
   "author": "Ghost Foundation",
   "license": "MIT",
+  "type": "module",
   "main": "server.js",
   "scripts": {
     "build": "vite build",
@@ -15,8 +16,8 @@
     "test:unit": "NODE_ENV=testing vitest run --coverage",
     "test": "yarn test:unit",
     "docker:test": "docker compose run --rm -it proxy-server yarn test:unit",
-    "lint:code": "eslint *.js *.ts src/ test/ --ext .js,.ts --cache",
-    "lint:code:fix": "eslint *.js *.ts src/ test/ --ext .js,.ts --cache --fix",
+    "lint:code": "eslint *.ts src/ test/ --ext .ts --cache",
+    "lint:code:fix": "eslint *.ts src/ test/ --ext .ts --cache --fix",
     "lint": "yarn lint:code",
     "preship": "yarn test",
     "ship": "STATUS=$(git status --porcelain); echo $STATUS; if [ -z \"$STATUS\" ]; then yarn version && git push --follow-tags; fi"
