+++ src/services/proxy/index.ts
@@ -1,1 +1,1 @@
-export * from './proxy';
+export * from './proxy.js';
